+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.exceptions.DukeException;
+import duke.utils.Storage;
+import duke.utils.TaskList;
+import duke.utils.Ui;
+
 public abstract class Command {
     public boolean isExit() {
         return false;
