+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.exceptions.DukeException;
+import duke.utils.Storage;
+import duke.utils.TaskList;
+import duke.utils.Ui;
+
 /**
  * Mark the task with the input index as done and prints a confirmation
  * message.
