+++ src/main/java/duke/tasks/Deadline.java
@@ -1,32 +1,33 @@
+package duke.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Optional;
 
-public class Event extends Task {
-    public final static String TASK_WORD = "event";
+public class Deadline extends Task {
+    public static final String TASK_WORD = "deadline";
+    private final LocalDateTime by;
 
-    private LocalDateTime at;
-
-    public Event(String description, boolean done, LocalDateTime at) {
+    public Deadline(String description, boolean done, LocalDateTime by) {
         super(description, done);
-        this.at = at;
+        this.by = by;
     }
 
     @Override
     public Optional<LocalDateTime> getTime() {
-        return Optional.of(this.at);
+        return Optional.of(this.by);
     }
 
     @Override
     public String getTaskType() {
-        return "Event";
+        return "duke.tasks.Deadline";
     }
 
     @Override
     public String toString() {
-        String checkbox = this.getDone() ? "[E][X]" : "[E][ ]";
+        String checkbox = this.getDone() ? "[D][X]" : "[D][ ]";
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd MMM yyyy, HH:mm");
-        String dateFormatted = "(at: " + this.at.format(formatter) + ")";
+        String dateFormatted = "(by: " + this.by.format(formatter) + ")";
         return checkbox + " " + super.getDescription() + " " + dateFormatted;
     }
 }
