+++ src/main/java/duke/tasks/Event.java
@@ -1,31 +1,34 @@
+package duke.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Optional;
 
-public class Deadline extends Task {
-    public static final String TASK_WORD = "deadline";
-    private final LocalDateTime by;
+public class Event extends Task {
+    public final static String TASK_WORD = "event";
 
-    public Deadline(String description, boolean done, LocalDateTime by) {
+    private LocalDateTime at;
+
+    public Event(String description, boolean done, LocalDateTime at) {
         super(description, done);
-        this.by = by;
+        this.at = at;
     }
 
     @Override
     public Optional<LocalDateTime> getTime() {
-        return Optional.of(this.by);
+        return Optional.of(this.at);
     }
 
     @Override
     public String getTaskType() {
-        return "Deadline";
+        return "duke.tasks.Event";
     }
 
     @Override
     public String toString() {
-        String checkbox = this.getDone() ? "[D][X]" : "[D][ ]";
+        String checkbox = this.getDone() ? "[E][X]" : "[E][ ]";
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd MMM yyyy, HH:mm");
-        String dateFormatted = "(by: " + this.by.format(formatter) + ")";
+        String dateFormatted = "(at: " + this.at.format(formatter) + ")";
         return checkbox + " " + super.getDescription() + " " + dateFormatted;
     }
 }
