+++ src/main/java/duke/utils/Parser.java
@@ -1,3 +1,19 @@
+package duke.utils;
+
+import duke.exceptions.DukeException;
+
+import duke.commands.Command;
+import duke.commands.AddCommand;
+import duke.commands.DeleteCommand;
+import duke.commands.MarkCommand;
+import duke.commands.UnmarkCommand;
+import duke.commands.ListCommand;
+import duke.commands.ExitCommand;
+import duke.commands.InvalidCommand;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Todo;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 
@@ -8,7 +24,7 @@ import java.time.format.DateTimeParseException;
  */
 
 public class Parser {
-    private static final InvalidCommand NUMBER_FORMAT = new InvalidCommand("Task index should be integers!");
+    private static final InvalidCommand NUMBER_FORMAT = new InvalidCommand("duke.tasks.Task index should be integers!");
 
     public Command parseCommand(String userInput) {
         String[] tokens = userInput.split(" ", 2);
